requirements.txt
ultralytics
streamlit
groq
python-dotenv
pillow